feat: Cria virtualização dos itens da tabela
