+++ lib/components/DataGrid/styles.ts
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+export const Container = styled.main`
+  display: flex;
+  flex-direction: column;
+
+  width: 100%;
+`;
