+++ lib/components/Table/index.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+import * as S from "./styles";
+
+type TableProps = {
+  children?: React.ReactNode;
+  style?: React.CSSProperties;
+};
+
+export const Table = ({ children, style }: TableProps) => {
+  return <S.Table style={style}>{children}</S.Table>;
+};
