+++ lib/components/Table/styles.ts
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const Table = styled.table`
+  border: ${({ theme }) => `1px solid ${theme.colors.gray200}`};
+  width: 100%;
+  table-layout: fixed;
+`;
