+++ lib/components/TableBody/styles.ts
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const TableBody = styled.tbody`
+  tr {
+    background-color: ${({ theme }) => theme.colors.white};
+  }
+`;
