+++ lib/components/TableFoot/index.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import * as S from "./styles";
+
+type TableFootProps = {
+  children?: React.ReactNode;
+};
+
+export const TableFoot = ({ children }: TableFootProps) => {
+  return <S.TableFoot>{children}</S.TableFoot>;
+};
