--- lib/components/atoms/Tbody/styles.ts
@@ -1,7 +1,0 @@
-import styled from "styled-components";
-
-export const Tbody = styled.tbody`
-  tr {
-    background-color: ${({ theme }) => theme.colors.white};
-  }
-`;
