--- lib/components/atoms/Tfoot/styles.ts
@@ -1,3 +1,0 @@
-import styled from "styled-components";
-
-export const Tfoot = styled.tfoot``;
