--- lib/components/organisms/DataGrid/styles.ts
@@ -1,7 +1,0 @@
-import styled from "styled-components";
-
-export const Scroll = styled.main`
-  overflow: auto;
-
-  height: 500px;
-`;
