+++ lib/helpers/get-reverse-direction-column-menu.ts
@@ -5,11 +5,7 @@ export const getReverseDirectionColumnMenu = (
   const lastColumnIndex = columnsLength - 1;
   const penultimateColumnIndex = columnsLength - 2;
 
-  if (
-    columnIndex === lastColumnIndex ||
-    columnIndex === penultimateColumnIndex
-  ) {
-    return true;
-  }
-  return false;
+  return (
+    columnIndex === lastColumnIndex || columnIndex === penultimateColumnIndex
+  );
 };
