+++ package.json
@@ -44,6 +44,7 @@
   "dependencies": {
     "@dnd-kit/core": "^6.3.1",
     "react-icons": "^5.4.0",
+    "react-virtuoso": "^4.12.5",
     "styled-components": "^6.1.13"
   }
 }
