--- src/models/mock-data-model.ts
@@ -1,8 +1,0 @@
-export interface MockDataModel {
-  id: number;
-  name: string;
-  phone: string;
-  cpf: string;
-  salary: number;
-  status: number;
-}
